data_extraction_requirements.txt
# Data Extraction API Requirements
# Core dependencies for the data extraction system

# Web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# HTTP requests and web scraping
requests==2.31.0
beautifulsoup4==4.12.2

# Git repository handling
GitPython==3.1.40

# Hugging Face integration
huggingface-hub==0.19.4

# Database
sqlite3  # Built into Python

# Scheduling
schedule==1.2.0

# Data processing
numpy==1.24.3
pandas==2.1.4

# Text processing and NLP
scikit-learn==1.3.2

# Async support
aiofiles==23.2.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1

# CORS middleware (included in FastAPI)
# python-multipart for file uploads (if needed later)
python-multipart==0.0.6

# Logging and monitoring
structlog==23.2.0

# Type hints
typing-extensions==4.8.0

# JSON handling (built into Python)
# hashlib (built into Python)
# pathlib (built into Python)
# concurrent.futures (built into Python)
# dataclasses (built into Python)

# Optional: for more advanced web scraping
selenium==4.15.2  # If needed for JavaScript-heavy pages
lxml==4.9.3       # Fast XML/HTML parser
